+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented here.
 
+## [Unreleased]
+
+### Added
+
+- Added `/nyslic` NYS Elevator License lookup for individual and business licenses with Adaptive Card search and results.
+
 ## [0.1.0] - 2026-06-24
 
 ### Added